演奏画面からリザルトへ移動する時の幕アニメーションをWindowsAnimationManagerを使用するよう変更。(ただし不完全の箇所あり) 譜面分岐のアニメーションをWindowsAnimationManagerを使用するよう変更(BranchAnime=TypeAのみ) WindowsAnimationManagerを使用するためのクラスを追加(DTXMania2から移植) xaファイルをサポート対象外とした 譜面分岐コースのenum追加 #64 演奏情報表示がクリア時演出より描画優先度が低くなっていたので修正 BMScroll時刻の取得時にPlayerが考慮されていなかったので修正 ポーズメニュー改良の準備

+++ DTXManiaプロジェクト/コード/ステージ/07.演奏/CAct演奏ゲージ共通.cs
@@ -383,7 +383,7 @@ namespace DTXMania
 		public void Damage( E楽器パート screenmode, E楽器パート part, E判定 e今回の判定, int player )
 		{
 			float fDamage;
-            int nコース = CDTXMania.stage演奏ドラム画面.n現在のコース[ player ];
+            int nコース = (int)CDTXMania.stage演奏ドラム画面.n現在のコース[ player ];
 
             switch ( e今回の判定 )
 			{
